requirements.txt
Django
mysqlclient
django-report-builder
django-datetime-widget
django-easy-pdf